fix: split client components for returns pages
